requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
colorama==0.4.4
distlib==0.3.2
Django==3.2.6
django-versatileimagefield==2.1
djangorestframework==3.12.4
drf-flex-fields==0.9.2
filelock==3.0.12
mysqlclient==2.0.3
Pillow==8.3.1
platformdirs==2.2.0
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.1
qrcode==7.2
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
